+++ Coding/ban_ve_xe_khach_web/src/main/java/com/cnwtt/ban_ve_xe_khach/repository/VeXeRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VeXeRepository extends JpaRepository<VeXe, Integer> {
+public interface VeXeRepository extends JpaRepository<VeXe, Long> {
     Page<VeXe> findAll(Pageable pageable);
 
 }
